+++ src/modules/health/health.module.ts
@@ -1,10 +1,8 @@
 import { HealthService } from "./health.service";
 import { Module } from "@nestjs/common";
 import { HealthController } from "./health.controller";
-import { UsersModule } from "../users/users.module";
-
 @Module({
-  imports: [UsersModule],
+  imports: [],
   controllers: [HealthController],
   providers: [HealthService],
   exports: [],